\name{bc_longlat_map_img_ras}
\alias{bc_longlat_map_img_ras}
\docType{data}
\title{
Raster image of Bolsa Chica for use with \code{\link{shark_vis_longlat}}
}
\description{
Raster image of Bolsa Chica for use with \code{\link{shark_vis_longlat}}
}
\usage{data("bc_longlat_map_img_ras")}
\format{
  The format is:
 'raster' chr [1:640, 1:640] "#A3CCFFFF" "#A3CCFFFF" "#A3CCFFFF" ...
}
\source{
Google Maps.
}
\examples{
data(bc_longlat_map_img_ras)
}
\keyword{datasets}
