\name{shark_data_longlat}
\alias{shark_data_longlat}
\docType{data}
\title{
Raw shark data spline-interpolated to 90-second intervals
}
\description{
Raw shark data spline-interpolated to 90-second intervals (matrix)
}
\usage{data(shark_data_longlat)}
\format{
  The format is:
 int [1:226400, 1:7] 1217951746 1217951836 1217951926 1217952016 1217952106 1217952196 1217952286 1217952376 1217952466 1217952556 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "date_as_sec" "lat" "lon" "t_intervals" ...
}
\source{
Espinoza, Mario, Farrugia, Thomas J., and Christopher G. Lowe. Habitat use, movements
and site fidelity of the gray smooth-hound shark in a newly restored Southern
California estuary. Journal of Experimental Marine Biology and Ecology, 2011.
}

\keyword{datasets}
